chore(locations): Moved locations code out of sdk and into separate file.
