+++ src/index.js
@@ -9,7 +9,11 @@ import * as modules from './module';
 import * as project from './project';
 import * as sdk from './sdk';
 
+import { getInstallPaths, locations } from './locations';
+
 export {
+	getInstallPaths,
+	locations,
 	modules,
 	project,
 	sdk
